requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
bcrypt==4.1.1
passlib==1.7.4
alembic==1.13.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pydantic-settings==2.1.0
aiosqlite==0.19.0
asyncpg==0.29.0
PyJWT
psycopg2-binary
openai>=1.0.0